# api/src/__generated__/openapi.yaml
openapi: 3.0.0
paths:
...
  /owners/{ownername}/projects/{project}:
    delete:
      description: Projectを削除します
      operationId: ProjectsController_remove
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: &ref_2
        - supabase-auth: []
      summary: ProjectsController_remove / Projectの削除
      tags: &ref_3
        - projects
  /owners/{ownername}/projects/{project}/access:
    get:
      description: 認証しているUserのProjectのアクセス権限を取得します
      operationId: ProjectsController_getAccess
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      role:
                        type: string
                        enum:
                          - admin
                          - editor
                          - viewer
                          - invited
                          - invalid
                    required:
                      - role
                required:
                  - statusCode
                  - message
                  - data
      security: *ref_2
      summary: ProjectsController_getAccess / Projectのアクセス権限の取得
      tags: *ref_3
    put:
      description: Projectのアクセス権限を更新します
      operationId: ProjectsController_updateAccess
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      requestBody:
        required: true
        description: 更新するアクセス権限の情報
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                role:
                  type: string
                  enum:
                    - admin
                    - editor
                    - viewer
                    - invited
                    - invalid
              required:
                - userId
                - role
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_2
      summary: ProjectsController_updateAccess / Projectのアクセス権限の更新
      tags: *ref_3
    delete:
      description: Projectのアクセス権限を削除します
      operationId: ProjectsController_removeAccess
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      requestBody:
        required: true
        description: アクセス削除するUserの情報
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
              required:
                - userId
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_2
      summary: ProjectsController_removeAccess / Projectのアクセス権限の削除
      tags: *ref_3
  /owners/{ownername}/projects/{project}/icon:
    get:
      description: Projectのアイコンを取得します
      operationId: ProjectsController_getIcon
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                description: プロジェクトアイコン取得レスポンス
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      iconUrl:
                        type: string
                    required:
                      - iconUrl
                required:
                  - statusCode
                  - message
                  - data
      security: *ref_2
      summary: ProjectsController_getIcon / Projectのアイコンの取得
      tags: *ref_3
    put:
      description: Projectのアイコンを更新します
      operationId: ProjectsController_updateIcon
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file: {}
              required:
                - file
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_2
      summary: ProjectsController_updateIcon / Projectのアイコンの更新
      tags: *ref_3
  /owners/{ownername}/projects/{project}/name:
    put:
      description: Project名を更新します
      operationId: ProjectsController_updateName
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      requestBody:
        required: true
        description: 更新するProject名
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_2
      summary: ProjectsController_updateName / Project名の更新
      tags: *ref_3
  /owners/{ownername}/projects/{project}/tags:
    post:
      description: Projectにタグを追加します
      operationId: ProjectsController_addTag
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      requestBody:
        required: true
        description: 追加するタグの情報
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
                description:
                  type: string
              required:
                - name
                - color
                - description
      responses:
        '200':
          description: 追加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_2
      summary: ProjectsController_addTag / Projectのタグの追加
      tags: *ref_3
  /owners/{ownername}/projects/{project}/tags/{tagId}:
    put:
      description: Projectのタグを更新します
      operationId: ProjectsController_updateTag
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
        - name: tagId
          required: true
          in: path
          description: タグID
          schema:
            type: string
      requestBody:
        required: true
        description: 更新するタグの情報
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
                description:
                  type: string
              required:
                - name
                - color
                - description
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_2
      summary: ProjectsController_updateTag / Projectのタグの更新
      tags: *ref_3
    delete:
      description: Projectのタグを削除します
      operationId: ProjectsController_removeTag
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
        - name: tagId
          required: true
          in: path
          description: タグID
          schema:
            type: string
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_2
      summary: ProjectsController_removeTag / Projectのタグの削除
      tags: *ref_3
  /owners/{ownername}/projects/{project}/settings/access:
    post:
      description: ProjectにUserを招待します
      operationId: ProjectsController_inviteUser
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
      requestBody:
        required: true
        description: 招待するUserの情報
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  type: string
                  format: email
              required:
                - identifier
      responses:
        '200':
          description: 招待成功
          content:
            application/json:
              schema:
                description: プロジェクト招待レスポンス
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      result:
                        type: string
                        enum:
                          - succeeded
                          - failed
                          - already invited
                    required:
                      - result
                required:
                  - statusCode
                  - message
                  - data
      security: *ref_2
      summary: ProjectsController_inviteUser / Projectへの招待
      tags: *ref_3
    get:
      description: Projectのロール付きメンバーを取得します
      operationId: ProjectsController_getUserWithRole
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            type: string
        - name: username
          required: true
          in: query
          description: User名
          schema:
            type: string
        - name: user_num
          required: false
          in: query
          description: User番号
          schema:
            type: number
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                description: プロジェクトユーザー一覧（ロール付き）
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        role:
                          type: string
                          enum:
                            - admin
                            - editor
                            - viewer
                            - invited
                            - invalid
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                        avatarUrl:
                          type: string
                      required:
                        - role
                        - id
                        - name
                        - avatarUrl
                required:
                  - statusCode
                  - message
                  - data
      security: *ref_2
      summary: ProjectsController_getUserWithRole / Projectのロール付きメンバーの取得
      tags: *ref_3
...
info:
  title: andTopic API
  description: andTopic API description
  version: '1.0'
  contact: {}
...
servers: []
...